Update CSCI446_design_doc_group21.tex
Each of these algorithm sections needs to have a design decisions section added too.  I am unsure what to put on this, and wanted to consult both of you first.  Should they each be fairly similar to the decisions stated in the sections for each of the other two algorithms?