Update CSCI446_final_report_group21.tex
I am referencing the figure from the example run report in my third paragraph, which I had above the paragraph, but I do not know how to add that in this document.  Also, I do not know the details of the experimental approach, so it still needs to be written.  I'm around from 9 to 12 tomorrow morning, if you want to meet.